feat: refactorization, resolves #21
