+++ src/consts/endpoints.go
@@ -1,6 +1,6 @@
 package consts
 
-import "github.com/eoussama/freego/core/models"
+import "github.com/eoussama/freego/src/models"
 
 var EndpointPing = models.Endpoint{Fragments: []interface{}{"ping"}}
 var EndpointGames = models.Endpoint{Fragments: []interface{}{"games"}}
