+++ src/enums/event.go
@@ -1,6 +1,6 @@
 package enums
 
-import "github.com/eoussama/freego/core/types"
+import "github.com/eoussama/freego/src/types"
 
 const (
 	EventFreeGames types.TEvent = "free_games"
