+++ src/types/announcementType.go
