+++ src/types/gameFlag.go
